WIP removed old tests and organized into layers
