+++ adapters/adapters_suite_test.go
@@ -1,0 +1,13 @@
+package adapters_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestAdapters(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Adapters Suite")
+}
