+++ usecases/get_story_test.go
@@ -1,10 +1,10 @@
-package storybranch_test
+package usecases_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	storybranch "github.com/carpeliam/git-story-branch"
+	usecases "github.com/carpeliam/git-story-branch/usecases"
 )
 
 type MockGitRepository struct {
@@ -17,10 +17,10 @@ func (mockGitRepo MockGitRepository) GetBranchName() string {
 type MockPivotalTrackerReader struct {
 }
 
-func (mockTrackerReader MockPivotalTrackerReader) GetStory(storyID int) *storybranch.Story {
+func (mockTrackerReader MockPivotalTrackerReader) GetStory(storyID int) *usecases.Story {
 	if storyID == 1234567890 {
-		return &storybranch.Story{
-			ID: "1234567890",
+		return &usecases.Story{
+			ID: 1234567890,
 			Description: "Description",
 			State: "delivered",
 		}
@@ -28,26 +28,12 @@ func (mockTrackerReader MockPivotalTrackerReader) GetStory(storyID int) *storybr
 	return nil
 }
 
-func (mockTrackerReader MockPivotalTrackerReader) GetStoryDescription(storyID int) string {
-	if storyID == 1234567890 {
-		return "Description"
-	}
-	return "Wrong Description"
-}
-
-func (mockTrackerReader MockPivotalTrackerReader) GetStoryState(storyID int) string {
-	if storyID == 1234567890 {
-		return "delivered"
-	}
-	return "unstarted"
-}
-
 var _ = Describe("Git Tracker name translator", func() {
 	It("should retrieve a Pivotal Tracker Story based on the current git branch name", func() {
 		mockGitRepo := MockGitRepository{}
 		mockTrackerReader := MockPivotalTrackerReader{}
 
-		story := storybranch.GetStory(mockGitRepo, mockTrackerReader)
+		story := usecases.GetStory(mockGitRepo, mockTrackerReader)
 
 		Expect(story.Description).To(Equal("Description"))
 		Expect(story.State).To(Equal("delivered"))
