+++ usecases/usecases_suite_test.go
@@ -1,0 +1,13 @@
+package usecases_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestUsecases(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Usecases Suite")
+}
